+++ selekt-java/src/main/kotlin/com/bloomberg/selekt/SQLConnection.kt
@@ -16,12 +16,13 @@
 
 package com.bloomberg.selekt
 
+import com.bloomberg.selekt.cache.LruCache
 import com.bloomberg.selekt.commons.forEachByPosition
 import com.bloomberg.selekt.commons.forUntil
 import javax.annotation.concurrent.NotThreadSafe
 
 @NotThreadSafe
-internal class SQLConnection constructor(
+internal class SQLConnection(
     path: String,
     private val sqlite: SQLite,
     private val configuration: DatabaseConfiguration,